--- dialects/mssql/mssql.go
@@ -1,127 +1,0 @@
-package mssql
-
-import (
-	"database/sql"
-	"fmt"
-	"regexp"
-	"strconv"
-
-	_ "github.com/denisenkom/go-mssqldb"
-	"gorm.io/gorm"
-	"gorm.io/gorm/callbacks"
-	"gorm.io/gorm/clause"
-	"gorm.io/gorm/logger"
-	"gorm.io/gorm/migrator"
-	"gorm.io/gorm/schema"
-)
-
-type Dialector struct {
-	DSN string
-}
-
-func (dialector Dialector) Name() string {
-	return "mssql"
-}
-
-func Open(dsn string) gorm.Dialector {
-	return &Dialector{DSN: dsn}
-}
-
-func (dialector Dialector) Initialize(db *gorm.DB) (err error) {
-	// register callbacks
-	callbacks.RegisterDefaultCallbacks(db, &callbacks.Config{})
-	db.Callback().Create().Replace("gorm:create", Create)
-	db.ConnPool, err = sql.Open("sqlserver", dialector.DSN)
-
-	for k, v := range dialector.ClauseBuilders() {
-		db.ClauseBuilders[k] = v
-	}
-	return
-}
-
-func (dialector Dialector) ClauseBuilders() map[string]clause.ClauseBuilder {
-	return map[string]clause.ClauseBuilder{
-		"LIMIT": func(c clause.Clause, builder clause.Builder) {
-			if limit, ok := c.Expression.(clause.Limit); ok {
-				if limit.Offset > 0 {
-					builder.WriteString("OFFSET ")
-					builder.WriteString(strconv.Itoa(limit.Offset))
-					builder.WriteString("ROWS")
-				}
-
-				if limit.Limit > 0 {
-					if limit.Offset == 0 {
-						builder.WriteString(" OFFSET 0 ROWS")
-					}
-					builder.WriteString(" FETCH NEXT ")
-					builder.WriteString(strconv.Itoa(limit.Limit))
-					builder.WriteString(" ROWS ONLY")
-				}
-			}
-		},
-	}
-}
-
-func (dialector Dialector) Migrator(db *gorm.DB) gorm.Migrator {
-	return Migrator{migrator.Migrator{Config: migrator.Config{
-		DB:                          db,
-		Dialector:                   dialector,
-		CreateIndexAfterCreateTable: true,
-	}}}
-}
-
-func (dialector Dialector) BindVarTo(writer clause.Writer, stmt *gorm.Statement, v interface{}) {
-	writer.WriteString("@p")
-	writer.WriteString(strconv.Itoa(len(stmt.Vars)))
-}
-
-func (dialector Dialector) QuoteTo(writer clause.Writer, str string) {
-	writer.WriteByte('"')
-	writer.WriteString(str)
-	writer.WriteByte('"')
-}
-
-var numericPlaceholder = regexp.MustCompile("@p(\\d+)")
-
-func (dialector Dialector) Explain(sql string, vars ...interface{}) string {
-	return logger.ExplainSQL(sql, numericPlaceholder, `'`, vars...)
-}
-
-func (dialector Dialector) DataTypeOf(field *schema.Field) string {
-	switch field.DataType {
-	case schema.Bool:
-		return "bit"
-	case schema.Int, schema.Uint:
-		var sqlType string
-		switch {
-		case field.Size < 16:
-			sqlType = "smallint"
-		case field.Size < 31:
-			sqlType = "int"
-		default:
-			sqlType = "bigint"
-		}
-
-		if field.AutoIncrement || field == field.Schema.PrioritizedPrimaryField {
-			return sqlType + " IDENTITY(1,1)"
-		}
-		return sqlType
-	case schema.Float:
-		return "float"
-	case schema.String:
-		size := field.Size
-		if field.PrimaryKey && size == 0 {
-			size = 256
-		}
-		if size > 0 && size <= 4000 {
-			return fmt.Sprintf("nvarchar(%d)", size)
-		}
-		return "nvarchar(MAX)"
-	case schema.Time:
-		return "datetimeoffset"
-	case schema.Bytes:
-		return "varbinary(MAX)"
-	}
-
-	return ""
-}
