--- dialects/postgres/postgres.go
@@ -1,102 +1,0 @@
-package postgres
-
-import (
-	"database/sql"
-	"fmt"
-	"regexp"
-	"strconv"
-
-	"gorm.io/gorm"
-	"gorm.io/gorm/callbacks"
-	"gorm.io/gorm/clause"
-	"gorm.io/gorm/logger"
-	"gorm.io/gorm/migrator"
-	"gorm.io/gorm/schema"
-	_ "github.com/lib/pq"
-)
-
-type Dialector struct {
-	DSN string
-}
-
-func Open(dsn string) gorm.Dialector {
-	return &Dialector{DSN: dsn}
-}
-
-func (dialector Dialector) Name() string {
-	return "postgres"
-}
-
-func (dialector Dialector) Initialize(db *gorm.DB) (err error) {
-	// register callbacks
-	callbacks.RegisterDefaultCallbacks(db, &callbacks.Config{
-		WithReturning: true,
-	})
-	db.ConnPool, err = sql.Open("postgres", dialector.DSN)
-	return
-}
-
-func (dialector Dialector) Migrator(db *gorm.DB) gorm.Migrator {
-	return Migrator{migrator.Migrator{Config: migrator.Config{
-		DB:                          db,
-		Dialector:                   dialector,
-		CreateIndexAfterCreateTable: true,
-	}}}
-}
-
-func (dialector Dialector) BindVarTo(writer clause.Writer, stmt *gorm.Statement, v interface{}) {
-	writer.WriteByte('$')
-	writer.WriteString(strconv.Itoa(len(stmt.Vars)))
-}
-
-func (dialector Dialector) QuoteTo(writer clause.Writer, str string) {
-	writer.WriteByte('"')
-	writer.WriteString(str)
-	writer.WriteByte('"')
-}
-
-var numericPlaceholder = regexp.MustCompile("\\$(\\d+)")
-
-func (dialector Dialector) Explain(sql string, vars ...interface{}) string {
-	return logger.ExplainSQL(sql, numericPlaceholder, `'`, vars...)
-}
-
-func (dialector Dialector) DataTypeOf(field *schema.Field) string {
-	switch field.DataType {
-	case schema.Bool:
-		return "boolean"
-	case schema.Int, schema.Uint:
-		if field.AutoIncrement || field == field.Schema.PrioritizedPrimaryField {
-			switch {
-			case field.Size < 16:
-				return "smallserial"
-			case field.Size < 31:
-				return "serial"
-			default:
-				return "bigserial"
-			}
-		} else {
-			switch {
-			case field.Size < 16:
-				return "smallint"
-			case field.Size < 31:
-				return "integer"
-			default:
-				return "bigint"
-			}
-		}
-	case schema.Float:
-		return "decimal"
-	case schema.String:
-		if field.Size > 0 {
-			return fmt.Sprintf("varchar(%d)", field.Size)
-		}
-		return "text"
-	case schema.Time:
-		return "timestamptz"
-	case schema.Bytes:
-		return "bytea"
-	}
-
-	return ""
-}
