+++ go.mod
@@ -3,12 +3,6 @@ module gorm.io/gorm
 go 1.14
 
 require (
-	github.com/denisenkom/go-mssqldb v0.0.0-20200428022330-06a60b6afbbc
-	github.com/erikstmartin/go-testdb v0.0.0-20160219214506-8d10e4a1bae5
-	github.com/go-sql-driver/mysql v1.5.0
 	github.com/jinzhu/inflection v1.0.0
 	github.com/jinzhu/now v1.1.1
-	github.com/lib/pq v1.1.1
-	github.com/mattn/go-sqlite3 v2.0.1+incompatible
-	gorm.io/gorm v1.9.12
 )
