+++ schema/model_test.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"gorm.io/gorm"
-	"gorm.io/gorm/tests"
+	"gorm.io/gorm/utils/tests"
 )
 
 type User struct {
