+++ tests/delete_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"gorm.io/gorm"
-	. "gorm.io/gorm/tests"
+	. "gorm.io/gorm/utils/tests"
 )
 
 func TestDelete(t *testing.T) {
