+++ tests/query_test.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"gorm.io/gorm"
-	. "gorm.io/gorm/tests"
+	. "gorm.io/gorm/utils/tests"
 )
 
 func TestFind(t *testing.T) {
