+++ tests/scanner_valuer_test.go
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	"gorm.io/gorm"
-	. "gorm.io/gorm/tests"
+	. "gorm.io/gorm/utils/tests"
 )
 
 func TestScannerValuer(t *testing.T) {
