+++ tests/scopes_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"gorm.io/gorm"
-	. "gorm.io/gorm/tests"
+	. "gorm.io/gorm/utils/tests"
 )
 
 func NameIn1And2(d *gorm.DB) *gorm.DB {
