+++ tests/transaction_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"gorm.io/gorm"
-	. "gorm.io/gorm/tests"
+	. "gorm.io/gorm/utils/tests"
 )
 
 func TestTransaction(t *testing.T) {
